# api/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Developer Portal (ADP) API
  version: 1.0.0-alpha
  description: |
    REST API for AI agent governance, audit trails, and documentation.
    Supports both PostgreSQL and SQLite backends.

    ## Overview

    ADP provides two core capabilities for AI coding agents:

    1. **Decision Audit Trail** -- every agent decision is logged with reasoning, confidence, and policy outcome.
    2. **Documentation Engine** -- auto-generates session summaries, risk reports, and pattern analysis from decision records.

    ## Response Formats

    The REST API (`adp-server`, port 8080) uses envelope wrappers for all responses:

    - **Single object**: `{"data": <object>}`
    - **Created object** (201): `{"data": <object>}`
    - **List**: `{"items": [...], "total": 10, "limit": 50, "offset": 0}`
    - **Error**: `{"error": "message", "code": "bad_request", "details": null}`

    ## Sidecar vs REST API

    The commit endpoints (`/v1/commits/*`) exist in two places with **different response formats**:

    - **REST API** (this spec, port 8080): Wraps responses in `{"data": {...}}` envelopes.
    - **MCP HTTP sidecar** (port 8081): Returns top-level JSON without envelopes. Git hooks expect the sidecar format.

    See the commit endpoint descriptions for details on both formats.

    ## Authentication

    ADP supports three authentication methods:

...

    The first registered user is automatically granted the admin role.
...

    ## Database Backends

...

...
servers:
...
paths:
...
  /v1/governance/check:
    post:
      operationId: checkAction
      summary: Check whether an action is allowed
      description: |
        Evaluates an action against all active policies (builtin, database, and Rego).
        Returns whether the action is allowed, requires approval, or is denied,
        along with matched policy names and any warnings.
      tags: [Governance]
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckActionRequest"
            example:
              session_id: "abc-123"
              trust_level: 2
              action_type: "file_write"
              target:
                path: "/etc/nginx/nginx.conf"
              context:
                reason: "Update proxy settings"
      responses:
        "200":
          description: Policy evaluation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/CheckActionResult"
              example:
                data:
                  allowed: false
                  requires_approval: true
                  denied_reasons:
                    - "Sensitive file modification requires trust level >= 4"
                  policy_names:
                    - "sensitive_files"
                  warnings:
                    - "File is in production configuration path"
                  restrictions: []
        "400":
          description: Missing session_id or action_type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "session_id is required"
                code: "bad_request"
                details: null
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/governance/approvals:
    post:
      operationId: requestApproval
      summary: Request human approval for a blocked action
      description: |
        Creates an escalation request when an action requires human approval.
        The request includes the action details, reason, and priority.
        Default priority is "normal". Requests expire based on escalation configuration.
      tags: [Governance]
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestApprovalRequest"
            example:
              session_id: "abc-123"
              action: "modify /etc/nginx/nginx.conf"
              action_type: "file_write"
              target:
                path: "/etc/nginx/nginx.conf"
              reason: "Need to update proxy timeout settings for performance improvement"
              policy_names:
                - "sensitive_files"
              priority: "high"
      responses:
        "201":
          description: Approval request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/EscalationRequest"
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      operationId: listApprovals
      summary: List approval requests
      description: Returns a paginated list of approval requests with optional filters.
      tags: [Governance]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: session_id
          in: query
          schema:
            type: string
          description: Filter by session ID
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, denied, expired]
          description: Filter by approval status
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, normal, high, critical]
          description: Filter by priority level
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paginated list of approval requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscalationListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/governance/approvals/pending:
    get:
      operationId: listPendingApprovals
      summary: List pending approval requests
      description: |
        Convenience endpoint that returns only approval requests with status "pending".
        Equivalent to `GET /v1/governance/approvals?status=pending`.
      tags: [Governance]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paginated list of pending approval requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscalationListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/governance/approvals/{id}:
    get:
      operationId: getApproval
      summary: Get an approval request by ID
      tags: [Governance]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/ApprovalId"
      responses:
        "200":
          description: Approval request details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/EscalationRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      operationId: resolveApproval
      summary: Resolve an approval request
      description: |
        Approves or denies a pending escalation request. Sets the approver comment
        and transitions the request to "approved" or "denied" status.
      tags: [Governance]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/ApprovalId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveApprovalRequest"
            example:
              approved: true
              comment: "Approved -- timeout change is safe for production"
      responses:
        "200":
          description: Approval resolved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/EscalationRequest"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # Audit -- Decisions
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT authentication via locally-generated tokens (POST /v1/auth/login)
        or external JWKS provider. Required in production (PostgreSQL mode).
        Optional in SQLite development mode.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Shared API key authentication. Set via ADP_API_KEY environment variable.

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    ApprovalId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Approval (escalation) request ID
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 500
      description: Maximum number of items to return
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Number of items to skip for pagination

  # ---------------------------------------------------------------------------
  # Reusable Responses
  # ---------------------------------------------------------------------------
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "not found"
            code: "not_found"
            details: null
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "authentication required"
            code: "unauthorized"
            details: null
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "internal server error"
            code: "internal_error"
            details: null

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
          enum:
            - bad_request
            - not_found
            - unauthorized
            - forbidden
            - internal_error
            - conflict
        details:
          description: Additional error context (nullable)
          oneOf:
            - type: object
            - type: "null"

    # -------------------------------------------------------------------------
    # Sessions
    # -------------------------------------------------------------------------
...
    CheckActionRequest:
      type: object
      required: [session_id, action_type]
      properties:
        session_id:
          type: string
          description: Active session ID
        trust_level:
          type: integer
          minimum: 1
          maximum: 5
          description: Trust level to evaluate against (overrides session default)
        action_type:
          type: string
          description: Type of action being checked (e.g., file_write, shell_execute, deploy)
        target:
          type: object
          additionalProperties: true
          description: Action target details (path, resource, etc.)
        context:
          type: object
          additionalProperties: true
          description: Additional context for policy evaluation
    CheckActionResult:
      type: object
      properties:
        allowed:
          type: boolean
          description: Whether the action is permitted
        requires_approval:
          type: boolean
          description: Whether human approval is required before proceeding
        denied_reasons:
          type: array
          items:
            type: string
          description: Reasons the action was denied (empty if allowed)
        policy_names:
          type: array
          items:
            type: string
          description: Names of policies that matched this action
        warnings:
          type: array
          items:
            type: string
          description: Non-blocking warnings about the action
        restrictions:
          type: array
          items:
            type: string
          description: Additional restrictions placed on the action
    RequestApprovalRequest:
      type: object
      required: [session_id, action, reason]
      properties:
        session_id:
          type: string
          description: Active session ID
        decision_id:
          type: string
          description: Related decision record ID (optional)
        action:
          type: string
          description: Human-readable description of the action
        action_type:
          type: string
          description: Action type category
        target:
          type: object
          additionalProperties: true
          description: Action target details
        reason:
          type: string
          description: Why the agent wants to perform this action
        policy_names:
          type: array
          items:
            type: string
          description: Policies that triggered the escalation
        context_summary:
          type: object
          additionalProperties: true
          description: Summary of the context for the approver
        priority:
          type: string
          enum: [low, normal, high, critical]
          default: "normal"
          description: Escalation priority
    ResolveApprovalRequest:
      type: object
      required: [approved]
      properties:
        approved:
          type: boolean
          description: Whether the action is approved
        comment:
          type: string
          description: Approver's comment or rationale
    EscalationRequest:
      type: object
      properties:
        id:
          type: string
          description: Unique escalation request ID
        session_id:
          type: string
        decision_id:
          type: string
          description: Related decision record ID
        action:
          type: string
          description: Action requiring approval
        action_type:
          type: string
        target:
          type: object
          additionalProperties: true
        reason:
          type: string
          description: Agent's reason for the action
        policy_names:
          type: array
          items:
            type: string
        context_summary:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [pending, approved, denied, expired]
        priority:
          type: string
          enum: [low, normal, high, critical]
        approver_id:
          type: string
          description: ID of the user who resolved the request
        approver_comment:
          type: string
          description: Comment from the approver
        requested_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
          description: When the request was approved or denied
    EscalationListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EscalationRequest"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # -------------------------------------------------------------------------
    # Audit -- Decisions
    # -------------------------------------------------------------------------
...